Remove use-hard-newlines as it causes problems
The problem appears when embedding source code in emails. You want the
paragraphs to refill automatically, but not the source code. I didn't
find any nice way to do it. I had to write this function but it is not
satisfactory because hard newlines are not visible when composing the
email:

(defun my:replace-soft-to-hard-newlines (region-start region-end)
  (interactive "r")
  (save-excursion
    (goto-char region-start)
    (while (and (<= (point) region-end)
                (re-search-forward "$" region-end t))
      (kill-line)
      (insert hard-newline))))